+++ plotnine/aes.py
@@ -8,13 +8,17 @@ from .utils import suppress
 
 __all__ = ['aes']
 
-
 all_aesthetics = {
     'alpha', 'angle', 'color', 'colour', 'fill', 'group', 'intercept',
     'label', 'lineheight', 'linetype', 'lower', 'middle', 'radius',
     'sample', 'shape', 'size', 'slope', 'stroke', 'upper', 'weight', 'x',
     'xend', 'xintercept', 'xmax', 'xmin', 'y', 'yend', 'yintercept',
     'ymax', 'ymin'}
+
+scaled_aesthetics = {
+    'x', 'y', 'alpha', 'color', 'colour', 'fill',
+    'linetype', 'shape', 'size', 'stroke'
+}
 
 CALCULATED_RE = re.compile(r'\.\.([a-zA-Z0-9_]+)\.\.')